+++ controllers/sleepinfo/sleepinfo_controller.go
@@ -128,7 +128,7 @@ func (r *SleepInfoReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 
 	logSecret := log.WithValues("secret", secretName)
 	if !resources.hasResources() {
-		if err = r.upsertSecret(ctx, log, now, secretName, req.Namespace, secret, sleepInfoData, resources); err != nil {
+		if err = r.upsertSecret(ctx, log, now, secretName, req.Namespace, sleepInfo, secret, sleepInfoData, resources); err != nil {
 			logSecret.Error(err, "fails to update secret")
 			return ctrl.Result{
 				Requeue: true,
@@ -149,7 +149,7 @@ func (r *SleepInfoReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		}, nil
 	}
 
-	if err = r.upsertSecret(ctx, log, now, secretName, req.Namespace, secret, sleepInfoData, resources); err != nil {
+	if err = r.upsertSecret(ctx, log, now, secretName, req.Namespace, sleepInfo, secret, sleepInfoData, resources); err != nil {
 		logSecret.Error(err, "fails to update secret")
 		return ctrl.Result{
 			Requeue: true,